requirements.txt
numpy
scipy
torch
torchaudio
soundfile
pydub
requests
openai
python-dotenv
uvloop
websockets
fastapi
uvicorn

faster-whisper==1.2.1
ctranslate2==4.4.0

silero-vad
accelerate
f5-tts @ git+https://github.com/SWivid/F5-TTS.git
